Working backwards solver!
If you have a fact "1+1=2" and a where rule "1+a=2"
then Calculator can now infer that a=1

The main change is that while the Matcher solves where rules,
it attempts to see if any equalties match an existing fact
In order to prevent variable collision, Mather first inlines variables
that have an existing definition, then rewrites existing variables with
unique non-char variable names.
After solving it maps back.

This recursive checking introduces the possiblity for stack-overflow, see comments.
A counter mitigate this for service health.