Nicer varaible rewrites in fact instead of test
When solving a+b=2, there is no conflict to resolve with 1+1=2.
But naively resolving with rule
a + b = c
| a + 1 = x
| y + 1 = b
| x + y = c

Would cause conflicts where 'a' is treated as being in the same scope.
Originally, the fix was to just rewrite all the variables in the original
expression 'a+b=2' to compiler-generated variables above 'z':
first '{' then '|' etc.

That is not very readable, so instead avoid rewriting any variable without a conflict.
When needing to generate, start from 'a'.

Also, instead of rewriting the current expressoin, rewrite the rule to i.e.
a + b = d | x + y = d

Also allow uppercase variables, and make variable generation aware of case.